combine work dispatching examples into one
