--- examples/work-dispatching-v3/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "work-dispatching-v3"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-env_logger = "0.8.4"
-retina-core = { path = "../../core" }
-retina-filtergen = { path = "../../filtergen" }
-retina-datatypes = { path = "../../datatypes" }
-lazy_static = "1.4.0"
-serde = { version = "1.0", features = ["derive"] }
-regex = "1.7.3"
-clap = { version = "3.2.23", features = ["derive"] }
